+++ src/main/kotlin/com/prasetia/erp/model/preventive/PreventiveCustomer.kt
@@ -10,17 +10,17 @@ class PreventiveCustomer(
         val tahun: String,
         val customer_name: String,
         val area: String,
-        val nilai_po: Long?,
-        val nilai_penagihan: Long?,
-        val nilai_budget: Long?,
-        val realisasi_budget: Long?,
-        val laba_rugi: Long?,
+        val nilai_po: Double?,
+        val nilai_penagihan: Double?,
+        val nilai_budget: Double?,
+        val realisasi_budget: Double?,
+        val laba_rugi: Double?,
         val customer_id:Long?,
-        val persent_penagihan:Long?,
-        val persent_budget:Long?,
-        val persent_laba_rugi:Long?,
+        val persent_penagihan:Double?,
+        val persent_budget:Double?,
+        val persent_laba_rugi:Double?,
         val area_id:Long?
 ){
-    constructor(): this(0, "", "", "", 0, 0, 0, 0,
-            0, 0, 0, 0, 0, 0)
+    constructor(): this(0, "", "", "", 0.0, 0.0, 0.0, 0.0,
+            0.0, 0, 0.0, 0.0, 0.0, 0)
 }
